@@ -10,9 +10,9 @@
   },
   "main": "src/gmusic-mini-player.js",
   "scripts": {
-    "build": "grunt build",
-    "test": "grunt test",
-    "lint": "grunt eslint"
+    "build": "gulp build",
+    "test": "npm run lint",
+    "lint": "eslint src"
   },
   "repository": {
     "type": "git",
@@ -23,27 +23,34 @@
     "url": "https://github.com/GPMDP/gmusic-mini-player.js/issues"
   },
   "dependencies": {
-    "brfs": "^1.4.2"
+    "brfs": "^1.4.2",
+    "gmusic.js": "^6.0.6"
   },
   "devDependencies": {
-    "babel-preset-es2015": "^6.6.0",
+    "babel-eslint": "^6.1.2",
+    "babel-plugin-transform-class-properties": "^6.10.2",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-register": "^6.9.0",
     "clean-css": "^3.4.8",
-    "eslint": "^2.9.0",
-    "eslint-config-airbnb-base": "^3.0.1",
-    "eslint-plugin-import": "^1.8.1",
-    "grunt": "^0.4.5",
-    "grunt-babel": "^6.0.0",
-    "grunt-browserify": "^4.0.1",
-    "grunt-cli": "^0.1.13",
-    "grunt-contrib-copy": "^0.8.2",
-    "grunt-contrib-uglify": "^0.11.0",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-eslint": "^18.1.0",
-    "grunt-execute": "^0.2.2",
-    "gruntify-eslint": "^1.3.1",
+    "eslint": "^3.1.0",
+    "eslint-config-airbnb-base": "^4.0.2",
+    "eslint-plugin-import": "^1.12.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-browserify": "^0.5.1",
+    "gulp-rename": "^1.2.2",
+    "gulp-uglify": "^1.5.4",
     "rework": "^1.0.1",
     "rework-comments": "^1.0.0",
     "rework-important": "0.0.1",
     "rework-namespace-css": "^0.1.4"
+  },
+  "babel": {
+    "presets": [
+      "es2015"
+    ],
+    "plugins": [
+      "transform-class-properties"
+    ]
   }
 }
